# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for generating traces and retrieving source code mappings for performance analysis",
        "title": "Tempo OTLP Trace Demo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.4.208:3202",
    "basePath": "/",
    "paths": {
        "/api/batch/process": {
            "post": {
                "description": "Processes a batch of items with comprehensive tracing. Generates 6-15 spans with 300-1500ms duration depending on batch size.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch"
                ],
                "summary": "Process a batch of items",
                "parameters": [
                    {
                        "description": "Batch processing request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Batch processed successfully",
                        "schema": {
                            "$ref": "#/definitions/models.BatchResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/order/create": {
            "post": {
                "description": "Creates an order with comprehensive tracing. Generates 10-12 spans with 600-1500ms duration. If sleep=true, adds 5s delay to payment processing to simulate slow operation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BatchRequest": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.BatchResponse": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "string"
                },
                "failed_count": {
                    "type": "integer"
                },
                "processed_count": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.OrderRequest": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sleep": {
                    "description": "If true, simulate slow operation by adding 5s delay to processPayment",
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.OrderResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "number"
                }
            }
        },
...
    }
}